♻️ refactor: separate type in Popover component
